fix: apply review recommendations - part 1
